+++ packages/uniforms-material/src/NumField.js
@@ -2,10 +2,11 @@ import React          from 'react';
 import TextField      from 'material-ui/TextField';
 import connectField   from 'uniforms/connectField';
 import filterDOMProps from 'uniforms/filterDOMProps';
+import {Component}    from 'react';
 
 const noneIfNaN = x => isNaN(x) ? undefined : x;
 
-const Num = ({
+const Num_ = ({
     decimal,
     disabled,
     errorMessage,
@@ -30,7 +31,7 @@ const Num = ({
         max={max}
         min={min}
         name={name}
-        onChange={(event, value) => onChange(noneIfNaN((decimal ? parseFloat : parseInt)(value)))}
+        onChange={onChange}
         ref={inputRef}
         step={decimal ? 0.01 : 1}
         type="number"
@@ -38,6 +39,39 @@ const Num = ({
         {...filterDOMProps(props)}
     />
 ;
+
+// TODO: Provide more generic solution.
+class Num extends Component {
+    constructor () {
+        super(...arguments);
+
+        this.state = {value: '' + this.props.value};
+
+        this.onChange = this.onChange.bind(this);
+    }
+
+    componentWillReceiveProps ({decimal, value}) {
+        const parse = decimal ? parseFloat : parseInt;
+
+        if (noneIfNaN(parse(value)) !== noneIfNaN(parse(this.state.value.replace(/[\.,]+$/, '')))) {
+            this.setState({value: value === undefined || value === '' ? '' : '' + value});
+        }
+    }
+
+    onChange ({target: {value}}) {
+        const change = value.replace(/[^\d\.,]/g, '');
+        const parsed = noneIfNaN((this.props.decimal ? parseFloat : parseInt)(change));
+
+        this.setState(
+            () => ({value: change}),
+            () => this.props.onChange(parsed)
+        );
+    }
+
+    render () {
+        return Num_({...this.props, onChange: this.onChange, value: this.state.value});
+    }
+}
 
 Num.defaultProps = {
     fullWidth: true