+++ packages/uniforms-semantic/src/NumField.js
@@ -2,10 +2,11 @@ import React          from 'react';
 import classnames     from 'classnames';
 import connectField   from 'uniforms/connectField';
 import filterDOMProps from 'uniforms/filterDOMProps';
+import {Component}    from 'react';
 
 const noneIfNaN = x => isNaN(x) ? undefined : x;
 
-const Num = ({
+const Num_ = ({
     className,
     decimal,
     disabled,
@@ -42,7 +43,7 @@ const Num = ({
                 max={max}
                 min={min}
                 name={name}
-                onChange={event => onChange(noneIfNaN((decimal ? parseFloat : parseInt)(event.target.value)))}
+                onChange={onChange}
                 placeholder={placeholder}
                 ref={inputRef}
                 step={step || (decimal ? 0.01 : 1)}
@@ -62,5 +63,38 @@ const Num = ({
         )}
     </div>
 ;
+
+// TODO: Provide more generic solution.
+class Num extends Component {
+    constructor () {
+        super(...arguments);
+
+        this.state = {value: '' + this.props.value};
+
+        this.onChange = this.onChange.bind(this);
+    }
+
+    componentWillReceiveProps ({decimal, value}) {
+        const parse = decimal ? parseFloat : parseInt;
+
+        if (noneIfNaN(parse(value)) !== noneIfNaN(parse(this.state.value.replace(/[\.,]+$/, '')))) {
+            this.setState({value: value === undefined || value === '' ? '' : '' + value});
+        }
+    }
+
+    onChange ({target: {value}}) {
+        const change = value.replace(/[^\d\.,]/g, '');
+        const parsed = noneIfNaN((this.props.decimal ? parseFloat : parseInt)(change));
+
+        this.setState(
+            () => ({value: change}),
+            () => this.props.onChange(parsed)
+        );
+    }
+
+    render () {
+        return Num_({...this.props, onChange: this.onChange, value: this.state.value});
+    }
+}
 
 export default connectField(Num);